Fix minor 32-bit

+++ vi_timing/test/test.cpp
@@ -89,7 +89,7 @@ int main()
 		"Information about the \'vi_timing\' library:" << "\n"
 		"\tBuild type: " << static_cast<const char*>(vi_tmInfo(VI_TM_INFO_BUILDTYPE)) << "\n"
 		"\tVer: " << reinterpret_cast<std::intptr_t>(VI_TM_INFO()) << "\n"
-		"\tBuild number: " << reinterpret_cast<std::intptr_t>(VI_TM_INFO(VI_TM_INFO_BUILDNUMBER)) << "\n"
+		"\tBuild number: " << static_cast<unsigned>(reinterpret_cast<std::intptr_t>(VI_TM_INFO(VI_TM_INFO_BUILDNUMBER))) << "\n"
 		"\tVersion: " << static_cast<const char*>(VI_TM_INFO(VI_TM_INFO_VERSION)) << "\n"
 		"\tCompile time: " << static_cast<const char*>(VI_TM_INFO(VI_TM_INFO_BUILDTIME)) << "\n" <<
 		std::endl;
